fix typo

+++ colorjson.go
@@ -150,7 +150,7 @@ func (f *Formatter) marshalValue(val interface{}, buf *bytes.Buffer, depth int) 
 	case int64:
 		buf.WriteString(f.sprintColor(f.NumberColor, strconv.FormatInt(v, 10)))
 	case float32:
-		buf.WriteString(f.sprintColor(f.NumberColor, strconv.FormatFloat(v, 'f', -1, 32)))
+		buf.WriteString(f.sprintColor(f.NumberColor, strconv.FormatFloat(float64(v), 'f', -1, 32)))
 	case float64:
 		buf.WriteString(f.sprintColor(f.NumberColor, strconv.FormatFloat(v, 'f', -1, 64)))
 	case bool:
